Refactored Options to record
